// frontend/src/openapi/models/UploadReserved.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Overbae API
 * Backend API for the Overclaw / Overmind platform.
 *
 * The version of the OpenAPI document: 2.0.0
 *
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { mapValues } from '../runtime';
/**
 *
 * @export
 * @interface UploadReserved
 */
export interface UploadReserved {
    /**
     *
     * @type {string}
     * @memberof UploadReserved
     */
    uploadId: string;
    /**
     *
     * @type {string}
     * @memberof UploadReserved
     */
    filename: string;
    /**
     *
     * @type {number}
     * @memberof UploadReserved
     */
    chunkBytes: number;
    /**
     *
     * @type {number}
     * @memberof UploadReserved
     */
    maxBytes: number;
}

/**
 * Check if a given object implements the UploadReserved interface.
 */
export function instanceOfUploadReserved(value: object): value is UploadReserved {
    if (!('uploadId' in value) || value['uploadId'] === undefined) return false;
    if (!('filename' in value) || value['filename'] === undefined) return false;
    if (!('chunkBytes' in value) || value['chunkBytes'] === undefined) return false;
    if (!('maxBytes' in value) || value['maxBytes'] === undefined) return false;
    return true;
}

export function UploadReservedFromJSON(json: any): UploadReserved {
    return UploadReservedFromJSONTyped(json, false);
}

export function UploadReservedFromJSONTyped(json: any, ignoreDiscriminator: boolean): UploadReserved {
    if (json == null) {
        return json;
    }
    return {

        'uploadId': json['upload_id'],
        'filename': json['filename'],
        'chunkBytes': json['chunk_bytes'],
        'maxBytes': json['max_bytes'],
    };
}

export function UploadReservedToJSON(json: any): UploadReserved {
    return UploadReservedToJSONTyped(json, false);
}

export function UploadReservedToJSONTyped(value?: UploadReserved | null, ignoreDiscriminator: boolean = false): any {
    if (value == null) {
        return value;
    }

    return {

        'upload_id': value['uploadId'],
        'filename': value['filename'],
        'chunk_bytes': value['chunkBytes'],
        'max_bytes': value['maxBytes'],
    };
}