// frontend/src/openapi/models/UploadState.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Overbae API
 * Backend API for the Overclaw / Overmind platform.
 *
 * The version of the OpenAPI document: 2.0.0
 *
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { mapValues } from '../runtime';
/**
 *
 * @export
 * @interface UploadState
 */
export interface UploadState {
    /**
     *
     * @type {string}
     * @memberof UploadState
     */
    uploadId: string;
    /**
     *
     * @type {number}
     * @memberof UploadState
     */
    received: number;
}

/**
 * Check if a given object implements the UploadState interface.
 */
export function instanceOfUploadState(value: object): value is UploadState {
    if (!('uploadId' in value) || value['uploadId'] === undefined) return false;
    if (!('received' in value) || value['received'] === undefined) return false;
    return true;
}

export function UploadStateFromJSON(json: any): UploadState {
    return UploadStateFromJSONTyped(json, false);
}

export function UploadStateFromJSONTyped(json: any, ignoreDiscriminator: boolean): UploadState {
    if (json == null) {
        return json;
    }
    return {

        'uploadId': json['upload_id'],
        'received': json['received'],
    };
}

export function UploadStateToJSON(json: any): UploadState {
    return UploadStateToJSONTyped(json, false);
}

export function UploadStateToJSONTyped(value?: UploadState | null, ignoreDiscriminator: boolean = false): any {
    if (value == null) {
        return value;
    }

    return {

        'upload_id': value['uploadId'],
        'received': value['received'],
    };
}